probleme

+++ src/modules/mod_admin/admin.php
@@ -12,15 +12,17 @@ if (isset($_POST['function'])) {
                 if($_SESSION['token'] ==($_POST['token'])){
                     $timestamp_ancien = time() - (15*60);
                     if($_SESSION['token_time'] >= $timestamp_ancien){
-                $id = $_POST['ID_adherent'];
-                echo "$id";
-                $req1 = Connexion::getConnexion()->prepare('DELETE from info_inscription where ID_adherent= ? ');
-                $req1->execute(array($id));
-                $req = Connexion::getConnexion()->prepare('DELETE from adherent where ID_adherent= ? ');
-                $req->execute(array($id));
-                
-                var_dump($id);
-                echo "Suppresion effectué";  
+                        if (isset($_POST["ID_adherent"])) {
+                            $id = $_POST['ID_adherent'];
+                            echo "$id";
+                            $req1 = Connexion::getConnexion()->prepare('DELETE from info_inscription where ID_adherent= ? ');
+                            $req1->execute(array($id));
+                            $req = Connexion::getConnexion()->prepare('DELETE from adherent where ID_adherent= ? ');
+                            $req->execute(array($id));
+                            
+                            var_dump($id);
+                            echo "Suppresion effectué";  
+                        }
                     }}}  
             break;
     }
